requirements.txt
flask==3.0.2
flask-restx==1.3.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
marshmallow==3.20.2 